requirements.txt
langchain
langchain_openai
langchain_core
pythin-dotenv